+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- mongo: Changed `MongoStore` to `Store` to be more idiomatic.
+- mongo: Changed `ConnectToMongo` to `Connect` to be more idiomatic.
+- mongo: Changed `NewDefaultMongoStore` to `NewDefaultStore` to be more idiomatic.
+
+### Fixed
+- documentation: typo in `user_manager` referring to clients instead of users.
 
 ## [v0.13.0-alpha] - 2018-06-14
 ### Added